+++ addons/website_livechat/static/src/legacy/public_livechat.js
@@ -2,7 +2,6 @@ odoo.define('website_livechat.legacy.website_livechat.livechat_request', functio
 "use strict";
 
 var utils = require('web.utils');
-var session = require('web.session');
 var LivechatButton = require('im_livechat.legacy.im_livechat.im_livechat').LivechatButton;
 
 
@@ -31,26 +30,6 @@ LivechatButton.include({
         // a button when the page is loaded.
         $(window).trigger('resize');
         return this._super(...arguments);
-    },
-
-    //--------------------------------------------------------------------------
-    // Handlers
-    //--------------------------------------------------------------------------
-
-    /**
-     * @override
-     * Called when the visitor closes the livechat chatter the first time (first click on X button)
-     * this will deactivate the mail_channel, clean the chat request if any
-     * and allow the operators to send the visitor a new chat request
-     */
-    _onCloseChatWindow: function (ev) {
-        this._super(ev);
-        var cookie = utils.get_cookie('im_livechat_session');
-        if (cookie) {
-            var channel = JSON.parse(cookie);
-            session.rpc('/im_livechat/visitor_leave_session', {uuid: channel.uuid});
-            utils.set_cookie('im_livechat_session', "", -1); // remove cookie
-        }
     },
 });
